Split parameters into multiple files
